Create  Human
In this implementation, we have a base class called CharacterClass that defines the common attributes for all the classes, such as base damage, defense, speed, and magic. We also have a constructor and getters to initialize and retrieve these attributes.

We then define five subclasses for each of the classes, and we override the useSkill function to provide unique abilities for each class. For example, the Fighter uses Double Strike, the Rogue uses Backstab, the Mage uses Fireball, the Acolyte uses Heal, and the Warrior uses Charge.

Note that this is just an example implementation and can be modified to fit the specific requirements of your game.